Break Algolia's custom process method into multiple steps
This will allow more flexibility in overriding parts of the processing
sequence for customized handling. In our case we need to render
multiple language versions of the site, and only push at the end once
all languages are rendered.

To that end we break the `process` method up into pre- and post-process
steps, and we break the `push` method up so that parts that need to run
on a per-language basis (specifically, applying hooks), and parts that
only need to run at the end once all records have been extracted and
collected.